unified infra for both panels
