+++ vsc-extension/src-views/input-panel/App.tsx
@@ -1,9 +1,16 @@
 import React from "react";
+import { RecoilRoot } from "recoil";
+import { ExtensionStateProvider } from "../shared/ExtensionState";
+import { InputPanel } from "../shared/InputPanel";
 
-type Props = {};
-
-const App = (props: Props) => {
-  return <div>This is the inputs panel</div>;
+const App = () => {
+  return (
+    <RecoilRoot>
+      <ExtensionStateProvider>
+        <InputPanel />
+      </ExtensionStateProvider>
+    </RecoilRoot>
+  );
 };
 
 export default App;
