+++ vsc-extension/src-views/shared/InputPanel.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+type Props = {};
+
+export const InputPanel = (props: Props) => {
+  return <div>InputPanel</div>;
+};
